requirements.txt
azure-functions==1.2.0
azure-storage-blob==1.5.0
azure-cosmosdb-table==1.0.5
Pillow==7.0.0